// client/src/context/create-project/types.ts
export enum UpdateProjectEnum {
  SS_JSON = "SS_JSON",
  SET_PROJECT = "SET_PROJECT",
  CLOSE_TEMPLATE = "CLOSE_TEMPLATE",
  UPDATE_MEDIA_LIST = "UPDATE_MEDIA_LIST",
  SELECT_TEMPLATE_HANDLER = "SELECT_TEMPLATE_HANDLER",
  CLICK_ON_SELECTED_TEMPLATE = "CLICK_ON_SELECTED_TEMPLATE",
  DELETE_SELECTED_TEMPLATE = "DELETE_SELECTED_TEMPLATE",
  ASSEMBLY_HANDLE_EVENT = "ASSEMBLY_HANDLE_EVENT",
  STAGEING_HANDLE_EVENT = "STAGEING_HANDLE_EVENT",
  TEMPLATE_HANDLE_EVENT = "TEMPLATE_HANDLE_EVENT",
  CANCEL_ACTIVE_TEMPLATEID = "CANCEL_ACTIVE_TEMPLATEID",
  HANDLE_SELECTION_CLEAR = "HANDLE_SELECTION_CLEAR",
  HANDLE_SELECTED_VIDEO_CLIP = "HANDLE_SELECTED_VIDEO_CLIP",
  HANDLE_EMPTY_VIDEOPLAYER_CLICKEVENT = "HANDLE_EMPTY_VIDEOPLAYER_CLICKEVENT",
  IS_JSONLOAD_LOADING = "IS_JSONLOAD_LOADING",
  IS_SSJSON_FINAL_MODAL = "IS_SSJSON_FINAL_MODAL",
  IS_JSONLOAD = "IS_JSONLOAD",
  HANDLE_UPDATE_PROJECTDATA = "HANDLE_UPDATE_PROJECTDATA",
  SET_SELECTED_VIDEO_CLIP = "SET_SELECTED_VIDEO_CLIP",
  SET_CLIP_PROJECT = "SET_CLIP_PROJECT",
  UPDATE_STAGE_FIELD = "UPDATE_STAGE_FIELD",
  SSJSON_FINALMODAL_IS_JSONLOAD = "SSJSON_FINALMODAL_IS_JSONLOAD",
  SSJSON_MODAL_DATA = "SSJSON_MODAL_DATA",
  SET_FINAL_VIDEO = "SET_FINAL_VIDEO",
  ADD_TEMPLATE_DATA = "ADD_TEMPLATE_DATA",
  GET_TEMPLATE_FIELD = "GET_TEMPLATE_FIELD",
  MEDIA_DELETE_FILE = "MEDIA_DELETE_FILE",
  UPDATE_ALBUM_VISIBILITY = "UPDATE_ALBUM_VISIBILITY",
  CLICK_ON_LIST_MEDIA = "CLICK_ON_LIST_MEDIA",
  UPDATE_MEDIA_TEMPLATE_FIELDS = "UPDATE_MEDIA_TEMPLATE_FIELDS",
  MOVE_MEDIA_TO_STAGE_FIELDS = "MOVE_MEDIA_TO_STAGE_FIELDS",
  SELETCED_MEDIA_CARD = "SELETCED_MEDIA_CARD",
  UPDATE_SUBCLIP = "UPDATE_SUBCLIP",
  UPDATE_STAGING_SUB_CLIP = "UPDATE_STAGING_SUB_CLIP",
  ON_DROP_REORDERING = "ON_DROP_REORDERING",
  UPDATE_STAGING_FIELD = "UPDATE_STAGING_FIELD",
  MOVE_MEDIA_TEMPLATE_FIELDS = "MOVE_MEDIA_TEMPLATE_FIELDS",
  DELETE_UPDATE_CLIP = "DELETE_UPDATE_CLIP",
  FINAL_VIDEOTO_MERGED = "FINAL_VIDEOTO_MERGED",
  FINAL_VIDEO_TOMERGED = "FINAL_VIDEO_TOMERGED",
  DELETE_FINAL_MERGED_VIDEO = "DELETE_FINAL_MERGED_VIDEO",
  FINAL_VIDEO_RENDERED = "FINAL_VIDEO_RENDERED",
  MODAL_RENDERED = "MODAL_RENDERED",
  ERROR_MODAL = "ERROR_MODAL",
  MEDIA_DROP_TO_FIELD_AND_DELETE = "MEDIA_DROP_TO_FIELD_AND_DELETE",
  SET_THEMES_ID = "SET_THEMES_ID",
  SIMILAR_TEMPLATE_CHANGE = "SIMILAR_TEMPLATE_CHANGE",
  UPDATE_TEMPLATE_STYLE = "UPDATE_TEMPLATE_STYLE",
  AUDIO_IMAGE_FIELD_EMPTY = "AUDIO_IMAGE_FIELD_EMPTY",
  AUDIO_IMAGE_TEMPLATE_FIELD_EMPTY = "AUDIO_IMAGE_TEMPLATE_FIELD_EMPTY",
  ENTER_NUMBER_TEMPLATE_FIELD = "ENTER_NUMBER_TEMPLATE_FIELD",
  SET_TEMPLATE_FIELD = "SET_TEMPLATE_FIELD",
  ADD_ALBUM_ID = "ADD_ALBUM_ID",
  HANDLE_CLIP_CURRENT_TIME = "HANDLE_CLIP_CURRENT_TIME",
  HANDEL_ALLOW_DRAG_DROP = "HANDEL_ALLOW_DRAG_DROP",
}